Create diverse files for editor
